Enhancement: Created the patient and expert patient immediately after sign up
Depending the choice for the 'role' parameter - whether the patient or the expert - the patient or expert profile is created automatically. Login and register request is made to /api/v1/login and /api/vi/register respectively. Login endoints returns username, firstname, lastname and the the while register endpoint returns the username. firstname, lastname and token